JS setup for flatpickr
