+++ app/javascript/packs/application.css
@@ -1,0 +1,1 @@
+@import 'flatpickr/dist/flatpickr.css';
